extract ColorSchemeProvider
