+++ gatsby-config.js
@@ -71,6 +71,7 @@ module.exports = {
           _reduxState: "src/_reduxState",
           svg: "src/svg",
           templates: "src/templates",
+          Themes: "src/Themes",
           utils: "src/utils",
         },
         extensions: [],
